Camera Scripts refactor
